refactor(package): 패키지 이동, 클래스명 일관성 통일
