+++ app/src/main/java/woowacourse/movie/view/moviemain/movielist/MovieListPresenter.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.view.moviemain.movielist
 
-import woowacourse.movie.data.movie.MoviePosterRepository
+import woowacourse.movie.domain.repository.MoviePosterRepository
 import woowacourse.movie.domain.repository.MovieRepository
 import woowacourse.movie.view.mapper.toUiModel
 import woowacourse.movie.view.model.MovieUiModel
