+++ app/src/main/java/woowacourse/movie/view/moviemain/setting/SettingFragment.kt
@@ -14,7 +14,7 @@ import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.google.android.material.switchmaterial.SwitchMaterial
 import woowacourse.movie.R
-import woowacourse.movie.data.setting.SettingPreferencesManager
+import woowacourse.movie.data.setting.SettingPreferencesRepository
 import woowacourse.movie.databinding.FragmentSettingBinding
 import woowacourse.movie.view.AlarmController
 import woowacourse.movie.view.model.ReservationUiModel
@@ -31,7 +31,7 @@ class SettingFragment : Fragment(), SettingContract.View {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        presenter = SettingPresenter(this, SettingPreferencesManager(requireContext()))
+        presenter = SettingPresenter(this, SettingPreferencesRepository(requireContext()))
         binding = DataBindingUtil.inflate(inflater, R.layout.fragment_setting, container, false)
         binding.presenter = presenter
         return binding.root
