+++ domain/src/main/java/woowacourse/movie/domain/repository/MoviePosterRepository.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.data.movie
+package woowacourse.movie.domain.repository
 
 interface MoviePosterRepository {
     fun findPoster(title: String): Int
